Nearest Target with a parallel job
Whether it’s better to use more total CPU time in exchange for shorter start-to-end time is highly contextually dependent. If we need to do other work at the same time, we might be better off restricting a job to run on fewer cores. In this example, we aren’t doing any other work, so the shorter start-to-end time is clearly preferable.